revert to old logic
