+++ src/transformations/add-storage-gaps.ts
@@ -1,12 +1,14 @@
-import { SourceUnit } from 'solidity-ast';
-import { findAll } from 'solidity-ast/utils';
+import { SourceUnit, ContractDefinition } from 'solidity-ast';
+import { findAll, isNodeType } from 'solidity-ast/utils';
 
 import { formatLines } from './utils/format-lines';
-import { Layout } from './utils/extract-layout';
+import { hasOverride } from '../utils/upgrades-overrides';
 import { getNodeBounds } from '../solc/ast-utils';
+import { StorageLayout } from '../solc/input-output';
 import { Transformation } from './type';
 import { TransformerTools } from '../transform';
 import { extractNatspec } from '../utils/extractNatspec';
+import { decodeTypeIdentifier } from '../utils/type-id';
 
 // By default, make the contract a total of 50 slots (storage + gap)
 const DEFAULT_SLOT_COUNT = 50;
@@ -24,36 +26,90 @@ export function* addStorageGaps(
         }
       }
 
-      const layout = Layout.fromContract(contract, getLayout(contract));
-      const gapSize = targetSlots - layout.moveToFreeSlot().getPosition().slot;
+      const gapSize = targetSlots - getContractSlotCount(contract, getLayout(contract));
 
-      if (gapSize <= 0) {
+      if (gapSize > 0) {
+        const contractBounds = getNodeBounds(contract);
+        const start = contractBounds.start + contractBounds.length - 1;
+
+        const text = formatLines(0, [
+          ``,
+          [
+            `/**`,
+            ` * @dev This empty reserved space is put in place to allow future versions to add new`,
+            ` * variables without shifting down storage in the inheritance chain.`,
+            ` * See https://docs.openzeppelin.com/contracts/4.x/upgradeable#storage_gaps`,
+            ` */`,
+            `uint256[${gapSize}] private __gap;`,
+          ],
+        ]);
+
+        yield {
+          kind: 'add-storage-gaps',
+          start,
+          length: 0,
+          text,
+        };
+      } else if (gapSize < 0) {
         throw new Error(
-          `Contract ${contract.name} uses more than the ${targetSlots} reserved slots.`,
+          `Contract ${contract.name} uses more then the ${targetSlots} reserved slots.`,
         );
       }
-
-      const contractBounds = getNodeBounds(contract);
-      const start = contractBounds.start + contractBounds.length - 1;
-
-      const text = formatLines(0, [
-        ``,
-        [
-          `/**`,
-          ` * @dev This empty reserved space is put in place to allow future versions to add new`,
-          ` * variables without shifting down storage in the inheritance chain.`,
-          ` * See https://docs.openzeppelin.com/contracts/4.x/upgradeable#storage_gaps`,
-          ` */`,
-          `uint256[${gapSize}] private __gap;`,
-        ],
-      ]);
-
-      yield {
-        kind: 'add-storage-gaps',
-        start,
-        length: 0,
-        text,
-      };
     }
   }
+}
+
+function getNumberOfBytesOfValueType(type: string) {
+  const details = type.match(/^t_(?<base>[a-z]+)(?<size>[\d]+)?$/);
+  switch (details?.groups?.base) {
+    case 'bool':
+    case 'byte':
+      return 1;
+    case 'address':
+      return 20;
+    case 'bytes':
+      return parseInt(details.groups.size, 10);
+    case 'int':
+    case 'uint':
+      return parseInt(details.groups.size, 10) / 8;
+    default:
+      throw new Error(`Unsupported value type: ${type}`);
+  }
+}
+
+function getContractSlotCount(contractNode: ContractDefinition, layout: StorageLayout): number {
+  // This tracks both slot and offset:
+  // - slot   = Math.floor(contractSizeInBytes / 32)
+  // - offset = contractSizeInBytes % 32
+  let contractSizeInBytes = 0;
+
+  // don't use `findAll` here, we don't want to go recursive
+  for (const varDecl of contractNode.nodes.filter(isNodeType('VariableDeclaration'))) {
+    if (varDecl.mutability === 'constant') {
+      continue;
+    }
+    if (varDecl.mutability === 'immutable' && hasOverride(varDecl, 'state-variable-immutable')) {
+      continue;
+    }
+
+    // try get type details
+    const typeIdentifier = decodeTypeIdentifier(varDecl.typeDescriptions.typeIdentifier ?? '');
+    const type = layout.types?.[typeIdentifier];
+
+    // size of current object from type details, or try to reconstruct it if
+    // they're not available try to reconstruct it, which can happen for
+    // immutable variables
+    const size = type
+      ? parseInt(type.numberOfBytes, 10)
+      : getNumberOfBytesOfValueType(typeIdentifier);
+
+    // used space in the current slot
+    const offset = contractSizeInBytes % 32;
+    // free space in the current slot (only if slot is dirty)
+    const free = (32 - offset) % 32;
+    // if the free space is not enough to fit the current object, then consume the free space to start at next slot
+    contractSizeInBytes += (size > free ? free : 0) + size;
+  }
+
+  return Math.ceil(contractSizeInBytes / 32);
 }
